fix: 修复编译错误和代码警告
- 修复 LOG_INFO/LOG_ERROR 中括号不匹配导致的宏调用错误
- 修复 std::cmp_greater 调用中多余的右括号
- 修复 std::ranges::remove 返回值类型问题，使用 .begin() 获取迭代器
- 使用 static_cast<size_t> 消除 narrowing conversion 警告
- 使用 std::string::starts_with 替代 substr 比较
- 使用 std::ranges::find 替代 std::find

+++ src/database.cpp
@@ -192,7 +192,7 @@ IPGeoService::IPGeoService(const std::string& city_db_path, const std::string& a
 
     LOG_INFO("IPGeoService initialized with cache size: " + std::to_string(cache_size)
              + ", shards: " + std::to_string(shard_count)
-             + ", max memory: " + std::to_string(max_memory_bytes / (1024 * 1024)) + "MB");
+             + ", max memory: " + std::to_string(max_memory_bytes / static_cast<size_t>(1024 * 1024)) + "MB");
 }
 
 LookupResult IPGeoService::lookup(const std::string& ip_address) const {
@@ -293,7 +293,7 @@ MACLookupService::MACLookupService(const std::string& oui_db_path, size_t cache_
 
     LOG_INFO("MACLookupService initialized with cache size: " + std::to_string(cache_size)
              + ", shards: " + std::to_string(shard_count)
-             + ", max memory: " + std::to_string(max_memory_bytes / (1024 * 1024)) + "MB");
+             + ", max memory: " + std::to_string(max_memory_bytes / static_cast<size_t>(1024 * 1024)) + "MB");
 }
 
 LookupResult MACLookupService::lookup(const std::string& mac_address) const {
